requirements.txt
anyio==3.7.1
build==0.10.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cryptocompare==0.7.6
cryptography==41.0.1
DateTime==5.1
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
msgpack==1.0.5
packaging==23.1
pip-tools==6.14.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.7.0
pyparsing==3.1.0
pyproject_hooks==1.0.0
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
zope.interface==6.0
